Parallel Module Loading: Refactor threads pool and architecture
Instead of creating new threads in each run, reuse same threads to load the modules and we can keep the utilization of threads and cpufreq higher.

And let subthreads dedicated to loading modules in parallel or alone, main thread only needs to update independent module pool. this architecture should be more simpler and readable.

Bug: 402263126
Change-Id: If3f07a18cd7c43132ff799d32786d2466f3110f1
Signed-off-by: Chungkai Mei <chungkai@google.com>