fastboot: Check USB configuration
USB device in it's descriptor may have multiple configurations. While it
is possible to change active configuration, Linux documentation advises
to avoid using USBDEVFS_SETCONFIGURATION for it. (source:
www.kernel.org/doc/html/latest/driver-api/usb/usb.html#synchronous-i-o-support)
Because of that, interfaces that we can freely use are those from
configuration that is currently active. The active configuration maybe
not the first one in the USB descriptor.

There is no known fastboot capable device that uses multiple USB
configurations, but to be on the safe side, application should check if
interface that will be used for fastboot comes from active USB
configuration.

This shouldn't be required on osx and windows, as these backends should
only offer interfaces that are available (i.e. are part of the active
USB configuration).

Bug: 370465224
Change-Id: If5c74bcfc5f9b2b2f0c0e01ad80ace900b92eaf4