fastboot: Add support for alternate interface
bNumInterfaces field from USB configuration descriptor tells about
number of interfaces not about number of interface descriptors within
that configuration. It is important to parse all interface descriptors
within configuration's wTotalLength, because one interface may have
multiple alternate settings. If we found matching fastboot interface
which is alternate setting, then we have to switch interface to that
setting to be able to use it.

Bug: 370465224
Change-Id: Ie16935b983291046b260d1bdeb57ceaa64091aa2