Revert "Revert "Add adb-remount-sh to presubmit & kernel-presubm..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2019f7d7-02ed-4f8e-98cd-038b36c579a7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibb7aabdd8971bb86abf2d2623aefb95e4d6a5bcd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com