Revert "Run module code only after KeyMints receive module info"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/53a1648b-6b67-4a23-8ed0-f706add533e0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388085199
Change-Id: I924e4b091f863e6c333e16387e0f12ed2bcf8be4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com