Revert "Run module code only after KeyMints receive module info"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0d137f24-8b23-4e62-b8f6-50f8b0c6ffca). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie34729060c4602534d891ce66a691a21f2352bbd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com