Revert "Add toolbox_ramdisk"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/20212a43-56c1-4cf7-bd4a-7eb5bf9f98dd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 389116741
Change-Id: I78d765b99337b9a32552d97b3149ada032d213dc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com