Revert "Ensure CancelUpdate() always works in recovery."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0cf1a61d-f05b-46b7-835f-b4eff149d1ef). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If5bad286c5c1c22cda8fae91ebce222fc60c1142
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com