Revert "libsnapshot: Set compile_multilib to "first""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5a78135b-7b8e-4e50-8ec1-1e0deb6faa53). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2e16fe7e5164c73db5bf3d22902de5cd09759bdc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com