Revert "bootstat: Add NTC PMIC shutdown reason."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/48dc4e12-09e6-4fa4-8c76-972ee99e6edd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I792a6ec95e2c1bb46d85576f16579a50a0b8f6b0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com