Revert "bootstat: Add NTC PMIC shutdown reason."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d619dd8f-be19-4e34-ad0d-3476bf682db5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8c7794052e78773b5930d74088a37479022af7c2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com