Revert "bootstat: Add NTC PMIC shutdown reason."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0d91b140-df72-40f1-bfee-0b1daf39817a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2e06d9fb6325bb43ec000c5969c4e96d52eb6165
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com