Revert "Revert "Revert^2 "modprobe: Always log failures to the k..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0d6831ac-6052-4a5e-abdf-2f8cc149b1cd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic4c9519717200aca09f1c34fea758d1b4cf5f2f4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com