Revert "bootstat: Add NTC PMIC shutdown reason."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/643181dc-fcf5-4137-b472-f6c5d397db77). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3087a25fb97d6e2c2e998dbe576e42a6df2d68fa
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com