Revert "bootstat: Add NTC PMIC shutdown reason."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb185859-d76f-4772-b8d0-fc44850886a6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idc39d4526dbf18b7730516181b4c632e93162bf4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com