Revert "bootstat: Add NTC PMIC shutdown reason."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/690a3e0f-4413-4695-a7f0-28276cd34832). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic640d5983803775e7d9c458f5870294f2a3119fe
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com