Revert "bootstat: Add NTC PMIC shutdown reason."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9e316e42-a7c5-40c8-9d75-9c01afe889c1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia2b02f623ea44cf300b41753400d5175fb41ec6b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com