Start bpfloader earlier in non-blocking manner to improve boot time
BPF programs from mainline modules are currently loaded by
`exec_start bpfloader`, but it blocks the rest of the boot sequence and
takes a long time. To improve boot time, this CL changes init to start
bpfloader with `start` (which doesn't block the subsequent operations)
at an earlier time as soon as apex is initialized, and to wait until
init.svc.bpfloader property becomes `stopped` before starting netd
(which needs to wait for bpfloader completion).

Bug: 376365651
Test: Check perfetto trace on boot
Test: Check that netd is not started until bpfloader finishes (by
      checking logcat and perfetto trace) even when bpfloader doesn't
      finish before zygote-start

Change-Id: Id3eadbdb475959a6e6d1d1291406e8533a8dd6b7