Revert "bootstat: Add NTC PMIC shutdown reason."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7acb9419-d1c7-4dd2-8984-2243b285cdb8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia190f4e50dbd9cf690bbc82d23d52818fe12ea47
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com