Revert "bootstat: Add NTC PMIC shutdown reason."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c144d149-f433-46b3-a938-f067ae95ed33). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I60037dfa6051a756fbf986733ff8e420223f2486
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com