Revert "[String8] Add string_view ctor"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8b41f33e-e574-49d4-83e4-d681a0c0fcac). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia81eb997ab5f2fae0fc436feb061f40f09a5fc02
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com