Revert "bootstat: Add NTC PMIC shutdown reason."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3e5250e0-ee05-4c6d-8757-87ce5e91187d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2d4eb02158dc53c8fe35f81fb68994ee57790b6f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com