Revert "bootstat: Add NTC PMIC shutdown reason."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a62569d-4e26-4a96-90a8-5f266387a2ef). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie0abcb7c17b17e59a3fd14cba7f6249bdf0484c9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com