Revert "bootstat: Add NTC PMIC shutdown reason."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1282ed06-d50e-4581-a02e-138e6b6a4378). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I437a79a38952e6775f593b0b159e40980d85738e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com