Revert "bootstat: Add NTC PMIC shutdown reason."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a18e955f-69ef-454b-b4bb-d364b308598f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5c9d936fa90bf3932c05efd8ecbaa9836ffcf377
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com