Revert "bootstat: refine sys.boot.reason if device boot up from ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ad84ba97-6e8c-40c8-b603-8d6401edcfb8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I01bd967a6e15da660e3dca1223b7bc70bbf0680c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com