Revert "libsnapshot: Increase the size of scratch to 2MB"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8fd88272-7979-4870-9377-d4b64395ef6d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5675fc0ee3ea0b59d76efb5aa96c46c4b78a2404
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com