Revert "trusty: keymint: trusty-keymint-apex.mk: fix condition"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/de6ccbfb-1ae3-4071-aecd-9f11f574cb37). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib045668412b2fd30452025970bec31fef6354e19
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com