Revert "trusty: keymint: trusty-keymint-apex.mk: fix condition"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a1cef148-2812-4b84-89d7-8b038be0e75b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifeaadbe567a632b0524f5d90ebf7e155f17ce530
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com