Revert "Remove space check from overlayfs test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e8c84bc4-97be-49f2-875f-124c67d695fc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id0ba9d12d459d593b42d138471b121715c74d082
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com