Revert "trusty: keymint: trusty-keymint-apex.mk: fix condition"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/458163a2-b716-4cc7-b48d-ed6c931fe4f3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I971a27080551fb879a82969e95cd214f063bd55c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com