Revert "Remove space check from overlayfs test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f924e43e-823c-4687-878d-9ffc5eba3155). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibba950e55b94317e6e1f2bdc627be4461c1abecf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com