Revert "Fix potential flaky debuggerd tests."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7e5b7cf1-4774-4cfa-92b7-dff028ad971d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib8489480342336e35e56ea658615a3c2f630df0c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com