Revert "libprefetch: Start prefetch service based on build"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5d574fcb-b957-4dbe-84ca-5a47f7c9400a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id73f5d37ee4fe2190a44607a3bbd097f7e12fb76
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com