Revert "bootstat: Add NTC PMIC shutdown reason."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/15d1448c-209c-4c64-98ba-5e0db674f114). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ica7df653681b759f62cac665a0d5c98059f48b2e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com