libprefetch: Start record/replay on trigger instead of property
Currently, prefetch `record` or `replay` command is executed
when the respective property is set, which is done by `start`
command. However, due to the timing of which property trigger
gets executed, this results in prefetch not starting neither
`record` or `replay` until the very end of boot phase.

Instead of going through `start` to determine which command
to run using property, we explicitly start `record` on `init`
trigger, and `replay` on `early-init` trigger respectively.

Additionally, instead of `start` command performing the check
to determine which command to run next, each command now will
perform their own check to determine whether they should proceed
or not.

Bug: 380766679
Test: Builds
Test: On cuttlefish, observe replay starting at early-init
and correctly exiting when replay condition is not met.
Test: On cuttlefish, observe record starting at init
and correctly exiting when record condition is not met.

Change-Id: Iad926450f00984d703f72c9c7b1876a0b3d3ecfd