Revert "[owners] Remove subrahmanyaman@google.com from gatekeepe..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1497d380-deb7-4d7a-bf7d-63df8723223f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia5a8ff38678dfe978168b88751b178a19ff0a6d1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com