Revert "libcutils: clean up the ashmem code a bit."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8d256ad3-9708-4d02-9ff2-781a29296144). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 401437287
Change-Id: I520ae3cb7245f795283111f2eac165f6a5183f4b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com