Revert "libcutils: clean up the ashmem code a bit."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/04c7609e-3a16-4651-afaf-5d17910e5c12). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 401388552
Change-Id: If7a1e5cc7288787474ab52a72e256689250a2dd3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com