Revert "libcutils: clean up the ashmem code a bit."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2a48cd0a-f6bd-483a-afd3-4fbc2609b633). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 401388552
Change-Id: Ic56ed38949db76fbd9b7593d7c72dddd615b785f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com