Revert "fs_mgr: Skip zram setup on swapon_all if mmd zram manage..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2527fd08-f9bd-48a4-bf5c-04d9c39b3d0b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I80f8c4c0bc47d501e491cc8530d787fba0d4f64a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com