Revert "init: Add a boot monitor for the second stage init"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2527fd08-f9bd-48a4-bf5c-04d9c39b3d0b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I50e11ad4f48194941793e3cdf8228aab44a2ef19
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com