Revert "trusty: utils: rpmb_dev: rpmb_dev.wv.system.rc: update i..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d2d5dc96-c4dd-4f07-9302-3d01edcf9107). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic975833aedeb76a949dc30fccb627900962a6372
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com