Revert "fs_mgr: Skip zram setup on swapon_all if mmd zram manage..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6d6b07bb-620a-4540-bfcd-b6cafde6b2d8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1140de3afef3cbfb129d8e5beca0f12736e344b7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com