Revert "init: Add a boot monitor for the second stage init"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6d6b07bb-620a-4540-bfcd-b6cafde6b2d8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6b457f66f0a58e26abda50a46b15bc520d5c3ca3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com