Revert "Cleanup python version properties"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5084bf7d-bf18-4ca7-9c0f-4fed5f0a5a4d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5ecc19bf723e5a6412abb610b8d5c1ea7a6c7d2f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com