Revert "Add init-mmd-prop.rc in case mmd is not built"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5952c5c0-0e65-424b-aeee-f79ca59300d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If6c06c9605f1ef09d6ae80f3e2764e0d7f9ed080
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com