Revert "Add init-mmd-prop.rc in case mmd is not built"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6d1bfc88-4649-4f7f-9a7f-7f4402dcd700). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4bc5792d407e5e80e6f8219c43a973d520c65188
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com