Revert "init: Add a boot monitor for the second stage init"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a59a5787-ff32-4ec4-b787-17aa019805ae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6b766d09b791582008f097ceed96d5bc3d11c63f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com