Revert "fs_mgr: Skip zram setup on swapon_all if mmd zram manage..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a59a5787-ff32-4ec4-b787-17aa019805ae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3741d22f95c5577beb2566737e703578babf5722
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com