Revert "Add init-mmd-prop.rc in case mmd is not built"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a9719e72-0c76-4b62-a633-925be924603e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic1242a8fbc625b3d79e11dfcf75003e72dfd6a0e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com