Revert "libprocessgroup: memcg v2 default MaxActivationDepth = 3"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/89733e71-2cc3-48c7-9acf-1c765c311afd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4881836536ba0e0be10f490594bf1823b98bd9b5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com