Revert "libprocessgroup: memcg v2 default MaxActivationDepth = 3"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/02109831-0785-4f74-8a77-67fa94e473c3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6b1f74289ecea46731540872cd6f6d9137c8cc2f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com