Revert "libprocessgroup: memcg v2 default MaxActivationDepth = 3"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4d10c90a-3581-45b7-8643-17a0cbddb355). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6ede205e6ba8a12e98d4a6afacfde6474688f634
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com