Revert "libprocessgroup: memcg v2 default MaxActivationDepth = 3"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e9e1a305-9705-4629-bbde-822d1400fe1b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic7d955b143247f8dd910dd3d6a7e02b4b98e2a7c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com