Revert "libprocessgroup: memcg v2 default MaxActivationDepth = 3"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0b3dbcf2-be34-40cd-aa8e-2ccd50f027ee). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 405493681
Change-Id: I3b373369544db13019e571653f7061684d5f0d6b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com