Revert "libprocessgroup: memcg v2 default MaxActivationDepth = 3"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b04ffdf6-f2ca-463b-bd5f-fe3f5374b309). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I98e448ba367cf5e3b89b070d2f222240c3a533fc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com