Revert "libprocessgroup: memcg v2 default MaxActivationDepth = 3"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4c4d0f7-a3a9-4d36-8d96-0de1917d9981). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib35614648abb9d17da3fe30c517d6cf629f46728
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com