Revert "libprocessgroup: memcg v2 default MaxActivationDepth = 3"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8b601cd7-7d6f-4a92-8a9b-1bcb8d27bd46). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia2822346d629dddd08aa42762121d1818ba036a0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com