Revert "Parallel Module Loading: Refactor threads pool and archi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/189c4ca5-2b32-4b48-a7d8-07334d1a1142). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6beb27b90f912043c401a92cea72eac61c920d44
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com