Revert "Parallel Module Loading: Refactor threads pool and archi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/08fef2cb-b9b9-4440-8c38-ddbbff46cfec). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I04ac02d6238d99d29ba285d0b24fba359a58877c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com