Revert "libprocessgroup: memcg v2 default MaxActivationDepth = 3"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cbd48e5e-97af-4ebe-9fbd-032919650f5b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I26a3de5c370ab6ba73c950f307a94a74a103d494
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com