Revert "Parallel Module Loading: Refactor threads pool and archi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6a03b2fb-5d53-43fe-a802-760cbdfa897b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia836c62ab5a96ac89de744caa13dbe1f0a99029a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com