Revert "Parallel Module Loading: Refactor threads pool and archi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9c6ae7db-00c0-4f25-b454-0c2c74cc885a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1c61f68a5e866445b4e8c06ffa6e2fbe06245138
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com