Revert "Parallel Module Loading: Refactor threads pool and archi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/837067d9-3b59-478f-882a-def2b34a176c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I91f9e1eb0f8228c3006268da229b51e2d02a0958
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com