Revert "Parallel Module Loading: Refactor threads pool and archi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dea4c577-4c9a-44fe-bff8-20c781c18f22). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I943f0d86e4b19559e96dec01648678fa8d82bc5a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com