Revert "Parallel Module Loading: Refactor threads pool and archi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c3765795-266c-4393-952e-e2ce6496b095). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie7523b64b6611557b50f1124d754d3ade50246b5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com