Revert "Parallel Module Loading: Refactor threads pool and archi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/39944e0f-01ec-4fe9-9f6b-8f2b32c3831c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I15e7ee9909028cb9750d9195d5a21818f0cc6049
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com