Revert "Parallel Module Loading: Refactor threads pool and archi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f7da76aa-8d69-4e69-a553-48d0d62a5c44). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5bf1918ef601823365e592e55c80f76c3b70c024
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com