Revert "Parallel Module Loading: Refactor threads pool and archi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f65f2880-fadf-47f0-976d-ca28b067a7ab). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7f301efa9cc3c290ffce543c787f00729091bb84
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com