Revert "Parallel Module Loading: Refactor threads pool and archi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/81917e6c-fba0-490c-9450-8a97901de15d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7668096748a79f243c89135b9e7e8b1df1bf84e9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com