Revert "Parallel Module Loading: Refactor threads pool and archi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ec6aee68-ebaa-474e-bb14-134c530dc2d7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I82420f839ba5690eb5705b0ddc2828c6a5360166
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com