Switch the keymint hal back to v3
Now keymint hal will pass the attestation version based on the
hal version to the keymint TA in TEE. When cfg "hal_v4" is
enabled, the attestation version passed to TA would be "400".
Default "libkmr_hal" and "libkmr_wire" library both have "hal_v4"
cfg enabled.

However, the keymint AIDL v4 is not frozen on android 15 QPR2
release, getInterfaceVersion() API will return the AIDL version
as "3" which is the latest stable AIDL version. Some vts tests
fail because of version mismatch as they will require the
"attestation version / 100" should be less than or equal to the
aidl version returned by getInterfaceVersion().

Switch the keymint hal back to v3 for android 15 as the keymint
AIDL v4 is not frozen.

Test: run vts -m VtsAidlKeyMintTargetTest

Change-Id: I39e200baa2d0ef6acc657a7ba36e7db993854d33
Signed-off-by: Ji Luo <ji.luo@nxp.com>