Switch KeyMint HAL service to v3 when frozen
When using the frozen HAL interfaces in Android 15, KeyMint is
(at most) version 3.  There are several places where this version
is visible.

Firstly, the standard places for an AIDL HAL have:

- The version of the HAL reported by `getInterfaceVersion()`: this
  generally gets set by linking to the appropriate generated code
  library from the AIDL definition, e.g.
  `android.hardware.security.keymint-V4-rust`.
- The vendor manifest vintf fragment declares the version, e.g. in
  `android.hardware.security.keymint-service.rust.trusty.xml`

For both of the above, the common infrastructure takes care of
reducing the version value by 1 when building in a configuration
that uses the previous frozen interfaces
(`RELEASE_AIDL_USE_UNFROZEN=false`).

However, KeyMint also exposes a version value in other places:

- A Package Manager flag gets populated with (100 * AIDL version).
  The existing `Android.bp` file deals with this by switching the
  XML file that defines the value, based on the
  `RELEASE_AIDL_USE_UNFROZEN` flag
- The KeyMint Rust reference implementation allows the HAL service
  to indicate the supported version to the TA code, which affects
  the attestation version field and the support of v4 features.

This CL addresses this latter point.

The most recent version of the reference implementation supports v4,
so also needs to be toggled back to v3 based on the
`RELEASE_AIDL_USE_UNFROZEN` flag.

Adapted from aosp/3566257

Test: VtsAidlKeyMintTargetTest
Change-Id: Ie8ffd70162021e5253daac90a6e36e431c2eeb2a