Revert "libprocessgroup: memcg v2 default MaxActivationDepth = 3"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8e0c5bb7-eb87-4947-9f62-6b1042a26c7e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iff2c02ebfa42ddfa42f9fadc6fc9dbe0168f23a7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com