Revert "create_snapshot: improve copy operation by reordering"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c32208fa-5ae5-45cd-9c1d-48d239dc4ad3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia7747b5dcb20570e7d23be710358f80fb6e0c5ca
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com