Check process uptime locally in the crashed process
Summary: Sometimes the crash service can't access the /proc/
dir of the process being dumped. Either due to permission issues or
because the process was killed inthe meantime, so we get no uptime information.

Record the uptime at the crashed process side and use that value
when writing the tombstone.

Test:   adb root
  adb shell "kill -35 `pidof vendor.oculus.hardware.composer-service`"
  adb pull /data/tombstones

Check that the tombstone has a correctly populated uptime field:
  Process uptime: 376s

Change-Id: I2fd35991d26e95a4867088c238cc714aaee659ee
Signed-off-by: Jarek Sobierski <jsobierski@fb.com>