fs_mgr_remount.cpp: fix the problem of failed remount
When we add two pieces of mounting information for a mount point in
the fstab. If the file systems of these two pieces of mounting
information are different, it may lead to a remount failure. For
example, if the following mounting table is used and the file system
of the vendor partition is erofs, the remount will fail:

vendor /vendor ext4  ro wait,slotselect,avb=vbmeta,logical,first_stage_mount
vendor /vendor erofs ro wait,slotselect,avb=vbmeta,logical,first_stage_mount

Test: adb root && adb remount
Change-Id: Ib2ae8700cfb0d31eb11f99617a8ff4eb87b8f2b4